Removed IBluetoothLE service and Simplified the Platform class structure
Moved platform classes into specific folders
