+++ DSoft.System.BluetoothLe/Extensions/Extensions.ios.mac.tvos.watchos.cs
