+++ DSoft.System.BluetoothLe/Platform/Android/BluetoothStatusBroadcastReceiver.android.cs
