+++ DSoft.System.BluetoothLe/Platform/Android/BondStatusBroadcastReceiver.android.cs
