+++ DSoft.System.BluetoothLe/Platform/Android/DefaultTrace.android.cs
