+++ DSoft.System.BluetoothLe/Platform/Android/GattCallback.android.cs
