+++ DSoft.System.BluetoothLe/Platform/Apple/BleCentralManagerDelegate.ios.mac.cs
