+++ DSoft.System.BluetoothLe/Platform/Apple/BleCentralManagerDelegate.tvos.watchos.cs
