+++ DSoft.System.BluetoothLe/Platform/Apple/DefaultTrace.ios.mac.tvos.watchos.cs
