+++ DSoft.System.BluetoothLe/Platform/Microsoft/BluetoothLE.uwp.netcore.netf.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 
 namespace System.BluetoothLe
 {
-    internal partial class BleImplementation
+    public partial class BluetoothLE
     {
         #region Fields
