+++ DSoft.System.BluetoothLe/Platform/Microsoft/DefaultTrace.uwp.cs
