+++ DSoft.System.BluetoothLe/Platform/Microsoft/ObservableBluetoothLEDevice.uwp.netcore.netf.cs
